fast_api/pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "facenet-pytorch>=2.5.3",
    "fastapi>=0.115.13",
    "numpy>=2.2.6",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "torch>=2.7.1",
    "uvicorn[standard]>=0.34.3",
    "websockets>=15.0.1",
]
